# src/test/resources/Collection_Service_API_Spec_v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "https://<host>:<port>/cp/collection-service/",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/file-gateway": {
            "get": {
                "tags": [
                    "file-gateway-controller"
                ],
                "summary": "Download the file",
                "operationId": "downloadFile",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filePath",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download a file from storage. If bucket name and path is not mentioned, the file will be searched under user uploaded folder.",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error downloading the file"
                    }
                }
            },
            "post": {
                "tags": [
                    "file-gateway-controller"
                ],
                "summary": "Upload a file",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "filePath",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "fileName": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "text/plain": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "fileName": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upload a file from file system",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error uploading the file"
                    }
                }
            },
            "delete": {
                "tags": [
                    "file-gateway-controller"
                ],
                "summary": "Delete a file",
                "operationId": "deleteFile",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filePath",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a file from storage. If bucket name and path is not mentioned, the file will be searched under user uploaded folder.",
                        "content": {
                            "text/plain": {}
                        }
                    },
                    "500": {
                        "description": "Error deleting the file"
                    }
                }
            }
        },
...
        "/api/v1/file-gateway/plan-file": {
            "get": {
                "tags": [
                    "file-gateway-controller"
                ],
                "summary": "Download the plan file",
                "operationId": "downloadPlanFile",
                "parameters": [
                    {
                        "name": "networkId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "collectorId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "collectorType",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TOPO_IGP",
                                "TOPO_BGPLS_XTC",
                                "TOPO_BGP",
                                "LSP_SNMP",
                                "LSP_PCEP_XTC",
                                "CONFIG_PARSE",
                                "TOPO_VPN",
                                "TRAFFIC_POLL",
                                "INVENTORY",
                                "TRAFFIC_DEMAND",
                                "LOGIN_FIND_MULTICAST",
                                "LOGIN_POLL_MULTICAST",
                                "SNMP_FIND_MULTICAST",
                                "SNMP_POLL_MULTICAST",
                                "NETFLOW",
                                "LAYOUT",
                                "EXTERNAL_SCRIPT",
                                "DARE",
                                "MULTICAST",
                                "DEMAND_MESH_CREATOR",
                                "DEMAND_FOR_LSPS",
                                "DEMAND_FOR_P2MP_LSPS",
                                "DEMAND_DEDUCTION",
                                "COPY_DEMANDS"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download a plan-file from storage. ",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error downloading the plan file"
                    }
                }
            }
        },
...
    },
...
}